Separate menu and dropdown (documentation still needs to be adjusted)
